+++ src/main/java/ladder/domain/result/Reward.java
@@ -1,4 +1,4 @@
-package ladder.domain.reward;
+package ladder.domain.result;
 
 import ladder.domain.player.Name;
