+++ src/test/java/ladder/domain/player/NameTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.player;
 
-import ladder.domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
