+++ src/test/java/ladder/domain/result/RewardTest.java
@@ -1,11 +1,8 @@
-package ladder.domain.reward;
+package ladder.domain.result;
 
-import ladder.domain.data.RewardData;
 import ladder.domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
